seperate code from design
