+++ src/components/chatbot/Rate.tsx
@@ -1,10 +1,10 @@
-import XIcon from "@/assets/x.svg";
 import { useEffect, useRef } from "react";
 import { Errors, Message } from "@/types";
 import { useRate } from "@/lib/useRate";
 import LikeIcon from "@/assets/like.svg"
 import PressedLikeIcon from "@/assets/pressed-like.svg"
-import "./footer.css"
+import XIcon from "@/assets/x.svg";
+import "./footer/footer.css"
 
 interface RateProps {
 	message: Message;
