+++ src/components/chatbot/chatbot.css
@@ -1,0 +1,50 @@
+.chatbot-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    touch-action: none;
+    pointer-events: auto;
+    z-index: 39;
+}
+
+.chatbot-popover-content {
+    direction: rtl;
+    margin: 0;
+    height: 100vh;
+    overflow: hidden;
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    z-index: 50;
+    padding: 0;
+    background-color: #ffffff;
+    display: flex;
+    flex-direction: column;
+}
+
+ .chatbot-popover-content.mobile {
+    margin-bottom: 0;
+    width: 100vw;
+    right: 0;
+    border-radius: 0;
+}
+
+.chatbot-popover-content.desktop {
+    width: 45vw;
+    right: auto;
+    border-radius: 8px;
+}
+
+.chatbot-popover-main {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    background-color: #f2f6fa;
+    overflow: hidden;
+}
