--- src/components/chatbot/footer.css
@@ -1,98 +1,0 @@
-
-.rating-form {
-    margin-top: 2px;
-    padding: 1rem;
-    border: 1px solid #4D6F8F;
-    background: white;
-    border-radius: 0.5rem;
-    color: black
-}
-
-.rate-border {
-    border-width: 0.25rem;
-    border-radius: 0.5rem;
-}
-
-.rate-text {
-    color: black;
-}
-
-.send-button-wrapper.disabled {
-    cursor: not-allowed;
-}
-
-
-.button-layout {
-    background-color: #2c5785;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 0.5rem;
-    padding: 0.5rem;
-    border-radius: 0.5rem;
-    color: white;
-        
-}
-
-.button-text.disabled {
-    cursor: not-allowed;
-}
-
-.button-layout.disabled {
-    opacity: 0.1;
-    cursor: not-allowed;
-}
-
-.button-send.disabled {
-    opacity: 0.1;
-    cursor: not-allowed;
-}
-
-
-/* .button-send.disabled {
-    border-radius: 0.25rem;
-    opacity: 0.1;
-    background: var(--Main, #2D5B85);
-} */
-
-
-.webiks-footer {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1%;
-    padding-left: 2%;
-    padding-right: 2%;
-
-
-    /* flex: 1 0 0; */
-    background-color: #E8EEF4;
-    /* height: "50px" */
-}
-
-
-
-.webiks-logo {
-    width: 3.75rem;
-    height: 1.45238rem;
-}
-
-.footer-disclaimer-text {
-    color: #4D6F8F;
-    text-align: right;
-    font-family: "Open Sans";
-    font-size: 0.7rem;
-    font-style: normal;
-    font-weight: 300;
-    line-height: normal;
-}
-
-.by-webiks {
-    display: flex;
-    flex-direction: row-reverse;
-    align-items: center;
-    gap: 0.5rem;
-    font-weight: 700;
-}
-
