--- src/components/chatbot/newQuestion.css
@@ -1,64 +1,0 @@
-.new-question-wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: linear-gradient(94deg, var(--Cyan, #3284FF) 11.95%, var(--Purple, #D65CFF) 83.06%);
-  border-radius: 54.376px;
-  border: none;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-  font-size: 1rem;
-  font-weight: 500;
-  height: 3.5rem;
-  position: relative;
-  padding-left: 0.25rem;
-  padding-right: 0.25rem;
-}
-
-.new-question-button-data {
-  display: flex;
-  width: 10rem;
-  direction: rtl;
-  justify-content: center;
-  align-items: center;
-  gap: 0.25rem;
-  background: white;
-  /* background: transparent; */
-  border-radius: 70px;
-  padding: 13px 21px;
-}
-
-.new-question-wrapper:hover .new-question-button-data {
-  background: transparent;
-}
-
-.ask-new-question-text {
-  font-family: inherit;
-  font-size: inherit;
-  font-weight: 700;
-  white-space: nowrap;
-  direction: rtl;
-  background: linear-gradient(94deg, #3284FF 11.95%, #D65CFF 83.06%);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-clip: text;
-}
-
-.new-question-wrapper:hover .ask-new-question-text {
-  background: white;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-clip: text;
-}
-
-.ask-new-question-icon {
-  width: 20px;
-  height: 20px;
-  flex-shrink: 0;
-  transition: filter 0.3s ease;
-}
-
-.new-question-wrapper:hover .ask-new-question-icon {
-  filter: brightness(0) invert(1);
-}
