+++ src/components/chatbot/newQuestion/newQuestion.css
@@ -1,0 +1,156 @@
+.new-question-section {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 16px;
+  padding: 24px 16px;
+  background: transparent;
+  width: 100%;
+  box-sizing: border-box;
+  font-family: "Open Sans", sans-serif;
+}
+
+.new-question-divider-container {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  gap: 16px;
+}
+
+.new-question-divider {
+  flex: 1;
+  height: 1px;
+  background: linear-gradient(90deg, transparent 0%, #E5E7EB 50%, transparent 100%);
+}
+
+.new-question-disclaimer {
+  font-size: 12px;
+  font-weight: 400;
+  line-height: 16px;
+  color: #6B7280;
+  text-align: center;
+  max-width: 400px;
+  font-family: "Open Sans", sans-serif;
+}
+
+.new-question-wrapper {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: linear-gradient(94deg, var(--Cyan, #3284FF) 11.95%, var(--Purple, #D65CFF) 83.06%);
+  border-radius: 54.376px;
+  border: none;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+  font-size: 1rem;
+  font-weight: 500;
+  height: 3.5rem;
+  position: relative;
+  padding-left: 0.25rem;
+  padding-right: 0.25rem;
+}
+
+.new-question-button-data {
+  display: flex;
+  width: 10rem;
+  direction: rtl;
+  justify-content: center;
+  align-items: center;
+  gap: 0.25rem;
+  background: #F2F6FA;
+  border-radius: 70px;
+  padding: 13px 21px;
+}
+
+.new-question-wrapper:hover .new-question-button-data {
+  background: transparent;
+}
+
+.ask-new-question-text {
+  font-family: inherit;
+  font-size: inherit;
+  font-weight: 700;
+  white-space: nowrap;
+  direction: rtl;
+  background: linear-gradient(94deg, #3284FF 11.95%, #D65CFF 83.06%);
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  background-clip: text;
+}
+
+.new-question-wrapper:hover .ask-new-question-text {
+  background: white;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  background-clip: text;
+}
+
+.ask-new-question-icon {
+  width: 20px;
+  height: 20px;
+  flex-shrink: 0;
+  transition: filter 0.3s ease;
+}
+
+.new-question-button {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding: 12px 20px;
+  background: transparent;
+  border: 1px solid #E5E7EB;
+  border-radius: 24px;
+  cursor: pointer;
+  transition: all 0.2s ease-in-out;
+  font-family: "Open Sans", sans-serif;
+  font-size: 14px;
+  font-weight: 500;
+  color: #374151;
+  white-space: nowrap;
+}
+
+.new-question-button:hover {
+  background: #F9FAFB;
+  border-color: #D1D5DB;
+  transform: translateY(-1px);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.new-question-button:active {
+  transform: translateY(0);
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
+}
+
+.new-question-button-text {
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 20px;
+  color: #374151;
+}
+
+.new-question-wrapper:hover .ask-new-question-icon {
+  filter: brightness(0) invert(1);
+}
+
+@media (max-width: 768px) {
+  .new-question-section {
+    padding: 16px;
+    gap: 12px;
+  }
+
+  .new-question-disclaimer {
+    font-size: 11px;
+    line-height: 14px;
+  }
+
+  
+  .new-question-button {
+    padding: 10px 16px;
+    font-size: 13px;
+  }
+  
+  .new-question-button-text {
+    font-size: 13px;
+  }
+}
