+++ src/components/chatbot/typingIndicator/typingIndicator.css
