+++ src/components/ui/input/input.css
@@ -1,0 +1,131 @@
+.chat-input-wrapper {
+    width: 100%;
+    position: relative;
+}
+
+.chat-input-field-container {
+    position: relative;
+    display: flex;
+    align-items: center;
+    width: 100%;
+    height: 3rem;
+    background: #FFFFFF;
+    border: 1px solid #E5E7EB;
+    border-radius: 9999px;
+    box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.1), 0px 1px 2px rgba(0, 0, 0, 0.06);
+    transition: all 0.2s ease-in-out;
+}
+
+.chat-input-field-container:focus-within {
+    border-color: #3B82F6;
+    box-shadow: 0px 0px 0px 3px rgba(59, 130, 246, 0.1), 0px 1px 3px rgba(0, 0, 0, 0.1);
+}
+
+.chat-input-field {
+    width: 100%;
+    height: 48px;
+    padding: 12px 24px;
+    border: none;
+    outline: none;
+    background: transparent;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 24px;
+    color: #111827;
+    font-family: "Open Sans", sans-serif;
+    border-radius: 9999px;
+    text-align: center;
+}
+
+.chat-input-field.with-submit-button {
+    padding-left: 64px;
+    padding-right: 24px;
+    text-align: right;
+}
+
+.chat-input-field::placeholder {
+    color: #9CA3AF;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 24px;
+}
+
+.chat-input-field:disabled {
+    cursor: not-allowed;
+    opacity: 0.5;
+    background-color: #F9FAFB;
+}
+
+.chat-input-submit-button {
+    position: absolute;
+    left: 12px;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 32px;
+    height: 32px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: transparent;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+    transition: all 0.2s ease-in-out;
+    padding: 0;
+}
+
+.chat-input-submit-button:hover {
+    background-color: rgba(59, 130, 246, 0.1);
+}
+
+.chat-input-submit-button:active {
+    transform: translateY(-50%) scale(0.95);
+}
+
+.chat-input-submit-button img {
+    width: 20px;
+    height: 20px;
+    display: block;
+}
+
+.chat-input-submit-button:disabled {
+  cursor: not-allowed;
+}
+
+.chat-input-error {
+    display: block;
+    margin-top: 8px;
+    padding: 8px 12px;
+    background-color: #FEF2F2;
+    border: 1px solid #FECACA;
+    border-radius: 6px;
+    color: #DC2626;
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 20px;
+    font-family: "Open Sans", sans-serif;
+}
+
+@media (max-width: 768px) {
+    .chat-input-field {
+        height: 44px;
+        padding: 10px 20px;
+        font-size: 16px;
+    }
+
+    .chat-input-field.with-submit-button {
+        padding-left: 56px;
+        padding-right: 20px;
+    }
+
+    .chat-input-submit-button {
+        left: 10px;
+        width: 28px;
+        height: 28px;
+    }
+
+    .chat-input-submit-button img {
+        width: 18px;
+        height: 18px;
+    }
+}
